dataacquisition/requirements.txt
paho-mqtt
pymongo
